fix(search): resolve floor names in area filters (#2235)
* fix(search): reject floor names as area filters

* fix(tools): expose typed bulk control operations

* fix(search): propagate branch tool errors

* fix(tools): reject obsolete bulk operation keys

* test: document changed regression helpers

* fix(tools): reject negative bulk timeouts

* fix(ci): preserve partial search errors

* fix(search): expand floor filters safely

* fix(bulk): preserve fail-soft operation handling

* fix(search): address floor resolution review feedback

* fix(bulk): preserve per-item validation and fractional timeouts

* fix(bulk): align schema and invalid-batch handling

* fix: harden timeout and registry edge cases

* fix: reject ambiguous floor and boolean inputs

* test: isolate legacy bulk validation path

* fix: reject infinite operation timeouts

* test: cover NaN timeout validation

+++ src/ha_mcp/utils/operation_manager.py
@@ -41,7 +41,7 @@ class DeviceOperation:
     expected_state: dict[str, Any] | None = None
     result_state: dict[str, Any] | None = None
     error_message: str | None = None
-    timeout_ms: int = 10000  # 10 second default timeout
+    timeout_ms: float = 10000  # 10 second default timeout
 
     @property
     def elapsed_ms(self) -> float:
@@ -84,7 +84,7 @@ class OperationManager:
         service_name: str,
         service_data: dict[str, Any],
         expected_state: dict[str, Any] | None = None,
-        timeout_ms: int = 10000,
+        timeout_ms: float = 10000,
     ) -> str:
         """Create a new device operation.
 
@@ -364,7 +364,7 @@ def store_pending_operation(
     service_name: str,
     service_data: dict[str, Any],
     expected_state: dict[str, Any] | None = None,
-    timeout_ms: int = 10000,
+    timeout_ms: float = 10000,
 ) -> str:
     """Store a new pending operation."""
     manager = get_operation_manager()
